large commit of working changes
