+++ src/OpenFec.py
@@ -139,42 +139,3 @@ class OpenFec:
             payload['page'] = page
             next_page = self.get_route(route, payload)
             yield next_page
-
-    def get_committee_by_id(self, committee_id: str, payload: dict) -> JSONType:
-        """gets committee info by committee_id
-            see https://api.open.fec.gov/developers/#/committee/get_committee__committee_id__
-
-        Args:
-            committee_id (str): A unique identifier assigned to each committee or filer registered with the FEC.
-                                In general committee id's begin with the letter C which is followed by eight digits.
-            payload (dict): request params object
-
-        Returns:
-            json: response as json object, of type CommitteeDetailPage
-        """
-        committee_id_without_quotes = committee_id.replace('"', '')
-        route = f'/committee/{committee_id_without_quotes}/'
-        url = self._get_route(route)
-        response = self._get_request(url, payload)
-        return response.json()
-
-    def get_committee_by_id_paginator(self, committee_id: str, payload={}) -> Generator:
-        """paginator to get committee info by committee_id
-            see https://api.open.fec.gov/developers/#/committee/get_committee__committee_id__
-
-        Args:
-            committee_id (str): A unique identifier assigned to each committee or filer registered with the FEC.
-                                In general committee id's begin with the letter C which is followed by eight digits.
-            payload (dict): request params object
-
-        Yields:
-            Generator: python Generator object to iteratate over get_committee_by_id
-        """
-
-        first_response = self.get_committee_by_id(committee_id, payload)
-        yield first_response
-        num_pages = first_response['pagination']['pages']
-        for page in range(2, num_pages + 1):
-            payload['page'] = page
-            next_page = self.get_committees(committee_id, payload)
-            yield next_page
